HW-1 Moved Class and Interface to Packages
