+++ class08/HudsonPark/ArchtectStudy/app/src/main/java/com/example/archstudy/data/repository/NaverQueryRepository.kt
@@ -1,0 +1,5 @@
+package com.example.archstudy.data.repository
+
+interface NaverQueryRepository {
+
+}
