+++ class08/HudsonPark/ArchtectStudy/app/src/main/java/com/example/archstudy/data/repository/NaverQueryRepositoryImpl.kt
@@ -1,0 +1,6 @@
+package com.example.archstudy.data.repository
+
+class NaverQueryRepositoryImpl {
+
+
+}
