+++ class08/HudsonPark/ArchtectStudy/app/src/main/java/com/example/archstudy/data/source/remote/NaverQueryRemoteDataSource.kt
@@ -1,0 +1,4 @@
+package com.example.archstudy.data.source.remote
+
+class NaverQueryRemoteDataSource {
+}
